Merge tracks by distance.
This seems to be buggy, though. Should move it to a branch.

brute is uselessly slow
kdtree is fast, but wrong (can probably be fixed)
dist is sometimes correct, but maybe not the right heuristic.
(kdetree could probably be merged in as a faster dist method).

Anyway, after the switch to saving in binary mode, tracks are much less
broken than before.